backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
python-dotenv
pydantic
pydantic[email]
pydantic-settings
langchain
langchain-community
langchain-google-genai
pillow
PyMuPDF
pgvector
sqlalchemy
psycopg2-binary
sentence-transformers
pandas
numpy
pytesseract